+++ src/bq_log/api/bq_log_api.cpp
@@ -207,7 +207,7 @@ namespace bq {
                     memcpy(thread_info_tls_.thread_name_, thread_name_tmp.c_str(), thread_info_tls_.thread_name_len_);
                 }
             }
-            uint32_t ext_info_length = sizeof(ext_log_entry_info_head) + thread_info_tls_.thread_name_len_;
+            uint32_t ext_info_length = static_cast<uint32_t>(sizeof(ext_log_entry_info_head) + thread_info_tls_.thread_name_len_);
             auto total_length = length + ext_info_length;
             auto epoch_ms = bq::platform::high_performance_epoch_ms();
